+++ test/types_test/num_f_test.dart
@@ -6,25 +6,25 @@ void main() {
   test('when num with NumF return true', () {
     /// *  num is int or double
     const json = {'age': 22.1, 'foo': 22};
-    final result = match(json, {'age': NumF, 'foo': NumF});
+    final result = matchMap(json, {'age': NumF, 'foo': NumF});
     expect(result, isTrue);
   });
   test('when null with NumF return false', () {
     /// *  num is int or double
     const json = {'age': null, 'foo': null};
-    final result = match(json, {'age': NumF, 'foo': NumF});
+    final result = matchMap(json, {'age': NumF, 'foo': NumF});
     expect(result, isFalse);
   });
   test('when num with NumOrNull return true', () {
     /// *  num is int or double
     const json = {'age': 22.1, 'foo': 22};
-    final result = match(json, {'age': NumOrNull, 'foo': NumOrNull});
+    final result = matchMap(json, {'age': NumOrNull, 'foo': NumOrNull});
     expect(result, isTrue);
   });
   test('when null with NumOrNull return true', () {
     /// *  num is int or double
     const json = {'age': null, 'foo': null};
-    final result = match(json, {'age': NumOrNull, 'foo': NumOrNull});
+    final result = matchMap(json, {'age': NumOrNull, 'foo': NumOrNull});
     expect(result, isTrue);
   });
 }