+++ test/types_test/string_f_test.dart
@@ -5,22 +5,22 @@ import 'package:test/test.dart';
 void main() {
   test('when String with StringF return true', () {
     const json = {'name': 'Ahmed Masoud'};
-    final result = match(json, {'name': StringF});
+    final result = matchMap(json, {'name': StringF});
     expect(result, isTrue);
   });
   test('when null with StringF return false', () {
     const json = {'name': null};
-    final result = match(json, {'name': StringF});
+    final result = matchMap(json, {'name': StringF});
     expect(result, isFalse);
   });
   test('when String with StringOrNull return true', () {
     const json = {'name': 'Ahmed Masoud'};
-    final result = match(json, {'name': StringOrNull});
+    final result = matchMap(json, {'name': StringOrNull});
     expect(result, isTrue);
   });
   test('when null with StringOrNull return true', () {
     const json = {'name': null};
-    final result = match(json, {'name': StringOrNull});
+    final result = matchMap(json, {'name': StringOrNull});
     expect(result, isTrue);
   });
 }