+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.3
+
+- Breaking Changes
+  - new `matchMap` , `matchF`
+  - new throwable parameter
+
 # 0.0.2
 
 - stable release